# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: >-
    A REST API for managing a mechanic shop. Customers can create accounts and
    service tickets, mechanics can manage assignments and inventory, and
    authenticated users can access resources based on their role. The API uses
    bearer JWT authentication.
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
...
securityDefinitions:
  customerBearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: >-
      Customer JWT issued by POST /customers/login. Send it as
      `Authorization: Bearer <token>`. Tokens use HS256, expire after one hour,
      and contain the role claim `customer`.
  mechanicBearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: >-
      Mechanic JWT issued by POST /mechanics/login. Send it as
      `Authorization: Bearer <token>`. Tokens use HS256, expire after one hour,
      and contain the role claim `mechanic`.
paths:
  /customers/:
    get:
      tags:
        - customers
      summary: "Get all customers"
      description: >-
        Return a paginated list of customer accounts. Responses are cached for
        60 seconds per unique query string.
      security: []
      parameters:
        - in: "query"
          name: "page"
          type: "integer"
          format: "int32"
          default: 1
          minimum: 1
          description: "The page number to return."
        - in: "query"
          name: "per_page"
          type: "integer"
          format: "int32"
          default: 10
          minimum: 1
          maximum: 100
          description: "The number of customers per page. Values above 100 are capped at 100."
      responses:
        "200":
          description: "Paginated list of customers"
          schema:
            $ref: "#/definitions/GetAllCustomersResponse"

    post:
      tags:
        - customers
      summary: "Create a customer"
      description: "Create a customer account with contact information and a password."
      security: []
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a customer"
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        "201":
          description: "Customer created successfully"
          schema:
            $ref: "#/definitions/CustomerResponse"
        "400":
          description: >-
            The email is already associated with an account, in which case the
            body is an ErrorResponse, or the payload failed validation, in which
            case the body is a ValidationErrorResponse.
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      tags:
        - customers
      summary: "Update the authenticated customer's account"
      description: >-
        Partially update the customer identified by the JWT subject. Only the
        fields present in the body are changed; a new password is re-hashed
        before it is stored.
      security:
        - customerBearer: []
      parameters:
        - in: "body"
          name: "body"
          description: "Customer fields to update. Every field is optional."
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        "200":
          description: "Customer updated successfully"
          schema:
            $ref: "#/definitions/CustomerResponse"
        "400":
          description: >-
            Missing, expired, invalid, or malformed bearer token, in which case
            the body is a MessageResponse, or invalid customer data, in which
            case the body is a ValidationErrorResponse.
          schema:
            $ref: "#/definitions/MessageResponse"
        "404":
          description: "Customer not found"
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      tags:
        - customers
      summary: "Delete the authenticated customer's account"
      description: "Delete the customer identified by the JWT subject."
      security:
        - customerBearer: []
      responses:
        "200":
          description: "Customer deleted successfully"
          schema:
            $ref: "#/definitions/MessageResponse"
        "400":
          description: "Missing, expired, or invalid bearer token"
          schema:
            $ref: "#/definitions/MessageResponse"
        "404":
          description: "Customer not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /customers/{customer_id}:
    get:
      tags:
        - customers
      summary: "Get a single customer"
      description: "Return one customer account by its ID."
      security: []
      parameters:
        - in: "path"
          name: "customer_id"
          type: "integer"
          format: "int32"
          required: true
          description: "The ID of the customer to return."
      responses:
        "200":
          description: "The requested customer"
          schema:
            $ref: "#/definitions/CustomerResponse"
        "404":
          description: "Customer not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /customers/login:
    post:
      tags:
        - customers
      summary: "Authenticate a customer"
      description: >-
        Authenticate a customer with an email address and password and return a
        customer JWT. Rate limited to 5 requests per minute.
      security: []
      parameters:
        - in: "body"
          name: "body"
          description: "Customer login credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        "200":
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
        "400":
          description: "The payload failed validation"
          schema:
            $ref: "#/definitions/ValidationErrorResponse"
        "401":
          description: "Invalid email or password"
          schema:
            $ref: "#/definitions/MessageResponse"
        "429":
          description: "Rate limit exceeded (more than 5 login attempts per minute)"
  /customers/my-tickets:
    get:
      tags:
        - customers
      summary: "Get the authenticated customer's service tickets"
      description: >-
        Return every service ticket belonging to the customer identified by the
        JWT subject, including the mechanics assigned and the parts used.
      security:
        - customerBearer: []
      responses:
        "200":
          description: "The customer's service tickets"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceTicketResponse"
        "400":
          description: "Missing, expired, or invalid bearer token"
          schema:
            $ref: "#/definitions/MessageResponse"
...
definitions:
  CustomerResponse:
    type: "object"
    description: "A customer account as returned by the API. The password is never included."
    required:
      - "id"
      - "name"
      - "email"
      - "phone"
    properties:
      id:
        type: "integer"
        format: "int32"
        description: "The customer's ID."
        example: 1
      name:
        type: "string"
        description: "The customer's full name."
        example: "Alex Morgan"
      email:
        type: "string"
        format: "email"
        description: "The customer's email address."
        example: "alex.morgan@example.com"
      phone:
        type: "string"
        description: "The customer's phone number."
        example: "555-0100"
  CreateCustomerPayload:
    type: "object"
    description: "Every field is required when creating a customer."
    required:
      - "name"
      - "email"
      - "phone"
      - "password"
    properties:
      name:
        type: "string"
        description: "The customer's full name."
        example: "Alex Morgan"
      email:
        type: "string"
        format: "email"
        description: "The customer's email address. Must be unique."
        example: "alex.morgan@example.com"
      phone:
        type: "string"
        description: "The customer's phone number. Must be unique."
        example: "555-0100"
      password:
        type: "string"
        format: "password"
        description: "The password for the customer account. Stored hashed."
        example: "secret123"
  UpdateCustomerPayload:
    type: "object"
    description: >-
      A partial update. Send only the fields you want to change; omitted fields
      are left untouched.
    properties:
      name:
        type: "string"
        description: "The customer's new full name."
        example: "Alex Morgan"
      email:
        type: "string"
        format: "email"
        description: "The customer's new email address. Must be unique."
        example: "alex.morgan@example.com"
      phone:
        type: "string"
        description: "The customer's new phone number. Must be unique."
        example: "555-0101"
      password:
        type: "string"
        format: "password"
        description: "A new password. Re-hashed before it is stored."
        example: "newsecret123"
  GetAllCustomersResponse:
    type: "object"
    required:
      - "customers"
      - "page"
      - "per_page"
      - "total_pages"
      - "total_customers"
    properties:
      customers:
        type: "array"
        description: "Customers included in the requested page."
        items:
          $ref: "#/definitions/CustomerResponse"
      page:
        type: "integer"
        format: "int32"
        description: "The current page number."
        example: 1
      per_page:
        type: "integer"
        format: "int32"
        description: "The number of customers requested per page."
        example: 10
      total_pages:
        type: "integer"
        format: "int32"
        description: "The total number of available pages."
        example: 3
      total_customers:
        type: "integer"
        format: "int32"
        description: "The total number of customers."
        example: 27
  LoginCredentials:
    type: "object"
    description: "Credentials accepted by both the customer and mechanic login endpoints."
    required:
      - "email"
      - "password"
    properties:
      email:
        type: "string"
        format: "email"
        description: "The account email address."
        example: "customer@example.com"
      password:
        type: "string"
        format: "password"
        description: "The account password."
        example: "secret123"
  LoginResponse:
    type: "object"
    required:
      - "status"
      - "message"
      - "token"
    properties:
      status:
        type: "string"
        description: "Login result status."
        example: "success"
      message:
        type: "string"
        description: "Result message for the login request."
        example: "successfully logged in."
      token:
        type: "string"
        description: "Bearer JWT to use for authenticated requests."
        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
  MessageResponse:
    type: "object"
    description: "A plain status message, used for deletions and auth failures."
    required:
      - "message"
    properties:
      message:
        type: "string"
        description: "A human-readable message describing the result."
        example: "Customer id: 1, successfully deleted."
  ErrorResponse:
    type: "object"
    description: "An error raised by the route itself, as opposed to schema validation."
    required:
      - "error"
    properties:
      error:
        type: "string"
        description: "A human-readable description of what went wrong."
        example: "Customer not found."
  ValidationErrorResponse:
    type: "object"
    description: >-
      Returned when the request body fails schema validation. Each key is a
      field name from the payload and each value is the list of problems found
      with that field.
    additionalProperties:
      type: "array"
      items:
        type: "string"
    example:
      email:
        - "Missing data for required field."
      phone:
        - "Not a valid string."
  MechanicResponse:
    type: "object"
    description: "A mechanic account as returned by the API. The password is never included."
    required:
      - "id"
      - "name"
      - "email"
      - "phone"
      - "salary"
    properties:
      id:
        type: "integer"
        format: "int32"
        description: "The mechanic's ID."
        example: 1
      name:
        type: "string"
        description: "The mechanic's full name."
        example: "Jordan Reyes"
      email:
        type: "string"
        format: "email"
        description: "The mechanic's email address."
        example: "jordan.reyes@example.com"
      phone:
        type: "string"
        description: "The mechanic's phone number."
        example: "555-0200"
      salary:
        type: "number"
        format: "double"
        description: "The mechanic's annual salary."
        example: 62000.0
  InventoryPartResponse:
    type: "object"
    description: "A part in the shop's catalog."
    required:
      - "id"
      - "name"
      - "price"
    properties:
      id:
        type: "integer"
        format: "int32"
        description: "The part's ID."
        example: 1
      name:
        type: "string"
        description: "The part's name."
        example: "Brake Pad Set"
      price:
        type: "number"
        format: "double"
        description: "The part's unit price."
        example: 89.99
  ServiceTicketPartResponse:
    type: "object"
    description: >-
      One part used on a service ticket, together with how many of that part
      were used.
    required:
      - "part"
      - "quantity"
    properties:
      part:
        $ref: "#/definitions/InventoryPartResponse"
      quantity:
        type: "integer"
        format: "int32"
        description: "How many of this part were used on the ticket."
        example: 2
  ServiceTicketResponse:
    type: "object"
    description: "A service ticket, including its assigned mechanics and the parts used."
    required:
      - "id"
      - "VIN"
      - "service_date"
      - "service_desc"
      - "customer_id"
      - "mechanics"
      - "parts"
    properties:
      id:
        type: "integer"
        format: "int32"
        description: "The service ticket's ID."
        example: 1
      VIN:
        type: "string"
        description: "The vehicle identification number of the car being serviced."
        example: "1HGCM82633A004352"
      service_date:
        type: "string"
        format: "date"
        description: "The date the service was performed, in YYYY-MM-DD format."
        example: "2025-03-14"
      service_desc:
        type: "string"
        description: "A description of the work done."
        example: "Replaced front brake pads and rotors."
      customer_id:
        type: "integer"
        format: "int32"
        description: "The ID of the customer who owns this ticket."
        example: 1
      mechanics:
        type: "array"
        description: "The mechanics currently assigned to this ticket."
        items:
          $ref: "#/definitions/MechanicResponse"
      parts:
        type: "array"
        description: "The parts used on this ticket and their quantities."
        items:
          $ref: "#/definitions/ServiceTicketPartResponse"
...